backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pymupdf
google-genai
requests
requests-oauthlib
python-dotenv
